requirements.txt
beautifulsoup4==4.12.3
countryinfo==0.1.2
deepl==1.18.0
folium==0.16.0
lxml==4.9.3
matplotlib==3.5.0
openai==0.28.0
opencv_python_headless==4.9.0.80
pandas==1.5.2
Pillow==10.4.0
playwright==1.45.1
pycountry==24.6.1
python-dotenv==1.0.1
pytz==2022.1
Requests==2.32.3
SocksiPy_branch==1.01
Telethon==1.35.0
tldextract==3.4.4
urllib3==1.26.5
wordcloud==1.9.2
